# demotape-example-api-only/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DemoTape Example API",
    "description": "A simple CRUD API to demonstrate DemoTape with backend-only apps.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "paths": {
    "/api/health": {
      "get": {
        "summary": "Health check",
        "tags": ["Health"],
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "ok" },
                    "uptime": { "type": "number", "example": 42.5 }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/items": {
      "get": {
        "summary": "List all items",
        "tags": ["Items"],
        "responses": {
          "200": {
            "description": "Array of items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Item" }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create an item",
        "tags": ["Items"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ItemInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Item created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Item" }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api/items/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": { "type": "integer" },
          "description": "Item ID"
        }
      ],
      "get": {
        "summary": "Get an item by ID",
        "tags": ["Items"],
        "responses": {
          "200": {
            "description": "The item",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Item" }
              }
            }
          },
          "404": {
            "description": "Item not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update an item",
        "tags": ["Items"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ItemInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Item updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Item" }
              }
            }
          },
          "404": {
            "description": "Item not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete an item",
        "tags": ["Items"],
        "responses": {
          "204": {
            "description": "Item deleted"
          },
          "404": {
            "description": "Item not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Item": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "name": { "type": "string", "example": "Widget A" },
          "description": { "type": "string", "example": "A standard widget" },
          "price": { "type": "number", "example": 9.99 }
        }
      },
      "ItemInput": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string", "example": "New Item" },
          "description": { "type": "string", "example": "A brand new item" },
          "price": { "type": "number", "example": 19.99 }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "Item not found" }
        }
      }
    }
  }
}